kontonummer in long umgebaut. mehrere Bugfixes

+++ Bankanwendung/src/g18/it1a/controller/ControllerUtils.java
@@ -7,15 +7,15 @@ import g18.it1a.model.Kunden;
 
 public class ControllerUtils {
 
-	public static Konto getKonto(int kontoNummer) throws AccountNotFoundException {
+	public static Konto getKonto(long kontoNummer) throws AccountNotFoundException {
 		int kundenNummer = getKundenNummer(kontoNummer);
 		Kunde kunde = Kunden.getKunde(kundenNummer);
 		return kunde.getKonto(kontoNummer);
 	}
 	
 	// TODO Try/Catch, Exceptions und Fehlerbehandlung
-	public static int getKundenNummer(int vonKonto) throws AccountNotFoundException {
-		String kontoNummer = Integer.toString(vonKonto);
+	public static int getKundenNummer(long vonKonto) throws AccountNotFoundException {
+		String kontoNummer = Long.toString(vonKonto);
 		int kontoNummerLength = kontoNummer.length();
 		if (kontoNummerLength < 6) {
 			throw new AccountNotFoundException();
